fix: rollback config file changes
