--- .idea/compiler.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="21">
-      <module name="travelin-android.feature.main" target="11" />
-      <module name="travelin-android.feature.onboarding.presentation.test" target="11" />
-      <module name="travelin-android.feature.test" target="11" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
